Add new All extension
Add a new extension to check that all values in an array match a given
condition. The All functionality leverages the Any functionality to avoid
duplicating code.
Add test for the new extension and update existing test to use a data
source instead of hard-coding test cases in the test method itself.