Add new methods for collapsing arrays
Add new extension methods for the single and multi-dimensional arrays to
collapse arrays down either to a single value or to a row/column.
Add methods on the one-based array interface and class that delegate down
to the array extensions.
Add new tests to ensure the new methods work as expected.