fix: random string generator producing duplicates
Update the `RandomStringGenerator` class to use a static `Random` variable
to ensure that when multiple instances of this class are created, they do
not all use the same initial seed (based on system time). By using a static
instance, the random generator is shared and will always use a single seed.
This issue has been observed on a project that uses this library, but was
not reproduceable as a test in this project.